# Bai1/github-copilot-vibe-coding-workshop/openapi.yaml
openapi: 3.0.1
info:
  title: Simple Social Media API
  description: A basic Social Networking Service (SNS) API that allows users to create, retrieve, update, and delete posts; add comments; and like/unlike posts.
  version: 1.0.0
  contact:
    name: Contoso Product Team
    email: support@contoso.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
  /posts:
    get:
      summary: List all posts
      description: Retrieve a list of all posts in reverse chronological order for browsing.
      operationId: listPosts
      tags:
        - Posts
      responses:
        '200':
          description: Successful response with list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new post
      description: Create a new post to share content with others.
      operationId: createPost
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{postId}:
    get:
      summary: Get a single post
      description: Retrieve details of a specific post by its ID.
      operationId: getPost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response with post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a post
      description: Update the content of an existing post.
      operationId: updatePost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a post
      description: Delete a specific post by its ID.
      operationId: deletePost
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          description: The ID of the post to delete
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Post:
      type: object
      description: A social media post
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the post
          example: 1
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Content of the post
          example: "Just had an amazing hiking experience in the mountains!"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the post was created
          example: "2025-05-30T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the post was last updated
          example: "2025-05-30T12:00:00Z"
        likesCount:
          type: integer
          description: Number of likes on the post
          example: 5
        commentsCount:
          type: integer
          description: Number of comments on the post
          example: 3
      required:
        - id
        - username
        - content
        - createdAt
        - updatedAt
        - likesCount
        - commentsCount
    CreatePostRequest:
      type: object
      description: Request payload for creating a new post
      properties:
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Content of the post
          example: "Just had an amazing hiking experience in the mountains!"
      required:
        - username
        - content
    UpdatePostRequest:
      type: object
      description: Request payload for updating an existing post
      properties:
        username:
          type: string
          description: Username of the post author
          example: "john_doe"
        content:
          type: string
          description: Updated content of the post
          example: "Updated: Just had an amazing hiking experience in the mountains with friends!"
      required:
        - username
        - content
...
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: "Post not found"
      required:
        - message